Refactor SourcesShowcase
