+++ app/components/windows/SourcesShowcase.vue.ts
@@ -8,7 +8,7 @@ import AddSourceInfo from './AddSourceInfo.vue';
 import { SourcesService, TSourceType, TPropertiesManager } from 'services/sources';
 import { ScenesService } from 'services/scenes';
 import { UserService } from 'services/user';
-import { WidgetsService, WidgetType } from 'services/widgets';
+import { WidgetsService, WidgetType, WidgetDisplayData } from 'services/widgets';
 
 
 type TInspectableSource = TSourceType | WidgetType | 'streamlabel';
@@ -73,12 +73,8 @@ export default class SourcesShowcase extends Vue {
     });
   }
 
-  widgetName(type: WidgetType) {
-    return this.widgetsService.getWidgetName(type);
-  }
-
-  widgetPlatform(type: WidgetType) {
-    return this.widgetsService.getWidgetPlatform(type);
+  widgetData(type: string) {
+    return WidgetDisplayData[this.widgetTypes[type]];
   }
 
   inspectedSource: TInspectableSource = null;
