+++ app/services/widgets/index.ts
@@ -1,0 +1,3 @@
+export * from './widgets-api';
+export * from './widgets';
+export * from './widgets-data';
